[INFRA] replace ssize_t with int

+++ test.h
@@ -125,7 +125,7 @@ TEST_API TEST_INLINE void set_console_color(int color) { (void)color; }
 
 TEST_API TEST_INLINE void test_print_string(char *str)
 {
-    ssize_t written_bytes = write(STDOUT_FILENO, str, (unsigned int)test_strlen(str));
+    int written_bytes = (int)write(STDOUT_FILENO, str, (unsigned int)test_strlen(str));
     (void)written_bytes;
 }
 
@@ -134,7 +134,7 @@ TEST_API TEST_INLINE void test_print_int(int val)
     char buf[12];
     char *p = buf + 11;
     unsigned int v = (unsigned int)val;
-    ssize_t written_bytes;
+    int written_bytes;
 
     *p = '\0';
 
@@ -153,7 +153,7 @@ TEST_API TEST_INLINE void test_print_int(int val)
     {
         *--p = '-';
     }
-    written_bytes = write(STDOUT_FILENO, p, (unsigned int)test_strlen(p));
+    written_bytes = (int)write(STDOUT_FILENO, p, (unsigned int)test_strlen(p));
     (void)written_bytes;
 }
